run all integration tests on all platforms
We currently exclude some integration tests on windows and macos.
This is (probably) done as an optimization to save some compute time
on these platform.  However we should have enough CI capacity to run
these tests everywhere, and it's better to spend that capacity, rather
than spend brain time to try and figure out what needs to be run cross
platform and what doesn't.